【修改说明】按检视意见修改告警

+++ services/media_engine/plugins/ffmpeg_adapter/common/reference_parser_manager.cpp
@@ -35,7 +35,7 @@ std::mutex ReferenceParserManager::mtx_;
 
 ReferenceParserManager::~ReferenceParserManager()
 {
-    if (referenceParser_) {
+    if (referenceParser_ != nullptr && destroyFunc_ != nullptr) {
         destroyFunc_(referenceParser_);
         referenceParser_ = nullptr;
     }
